Game
Copied and paste some code (Not from Chat GPT) from github to add to the file and link them to the index that will send you to 3 different games depending on what you click. Also changed some code in the game.html of renaming some statements that insult you depending on the results and clicking too early.